Edited tutor test cases by merging some of the student and teacher test cases together
